Add support for -x, -y and -width
Specifying the -x or -y argument allows you to control the position of
where dmenu is rendered on the focused monitor, unless the given
positions would cause dmenu to be rendered off screen. When -x is
specified, the width is automatically adjusted to not be rendered off
screen.

Specifying -width allows you to control the width of the rendered dmenu,
unless the specified width is too large for the focused monitor. The
height of dmenu can already be controlled through the -l argument